requirements.txt
Flask
opencv-python
numpy
pandas
matplotlib
tensorflow
Pillow
torch
torchvision